Refactor ChatBox
