+++ communication-service/socket-controller.js
@@ -1,9 +1,11 @@
 const createEventListeners = (socket, io) => {
   socket.on('message', (data) => {
     io.to(data.room).emit('messageResponse', data);
+    io.to(data.room).emit('stop-typingMessage');
   });
 
   socket.on('typing', (data) => {
+    console.log('typing');
     io.to(data.room).emit('typingMessage', data);
   });
