+++ frontend/src/components/room/chat/ChatBubble.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { Box, Typography } from '@mui/material';
 import { ChatModel } from './chat-model';
 
-const Message = (props: { chat: ChatModel; isSelf: boolean }) => {
+const ChatBubble = (props: { chat: ChatModel; isSelf: boolean }) => {
   const { chat, isSelf } = props;
 
   return (
@@ -24,4 +24,4 @@ const Message = (props: { chat: ChatModel; isSelf: boolean }) => {
   );
 };
 
-export default Message;
+export default ChatBubble;
